workflow_examples/requirements.txt
ray[train,tune,serve,data] == 2.54.0
fastapi[all] == 0.114.1
keras == 2.15.0
mlflow == 3.2.0
flytekit==1.15.4
pandas <= 2.1.4
tensorflow
pillow
s3fs
kubernetes
prometheus-api-client == 0.5.5
